Update next-steps.md
Added recommendations for some links to replace currently broken ones. Because I wanted you to review the resources and consider their validity, I did not actually update the text here, except to add a comment with the links. Ultimately, that comment should be deleted, the broken links removed, and the new links incorporated.