Update _index.md
Earthquake trend story link and title updated. While you can go to this page on Tableau Public, it doesn't render any data. If you search their site, you can find a similar article with a slightly different title that does render. I switched the title and link to be this one (https://public.tableau.com/app/profile/tableau.docs.team/viz/EarthquakeTrendStory2/Finished-Earthquakestory), but it may be worth double checking that this is a good example since it's not what the original linked to.